Update WordSearchPuzzle.java
i added in the spaceForWord method which is a helper method in order to make the actual puzzle. it takes in the word, a row, a column and a direction and will return true or false if there is space for the word to go in. It also allows for interlocking of words.
e.g.

there
a    a
n    c
k    h

(see how "tank" and "each" use letters from the word "there".